+++ packages/webapp/src/webview/ui/components/App/App.tsx
@@ -1,5 +1,5 @@
 import type { ReactElement } from 'react';
-import React from 'react';
+import React, { useEffect } from 'react';
 import { useSelector } from 'react-redux';
 import { VSCodeProgressRing } from '@vscode/webview-ui-toolkit/react';
 import { AppState } from '../../../types';
@@ -11,6 +11,7 @@ import { FiltersRibbon } from '../Header/Filters';
 import { FocusZone, FocusZoneDirection } from '@fluentui/react-focus';
 import './App.scss';
 import { initIcons } from '../UIComponentsLib/Icons';
+import InfiniteScroll from 'react-infinite-scroll-component';
 
 initIcons();
 
@@ -21,6 +22,46 @@ initIcons();
  */
 export function App(): ReactElement {
     const appState = useSelector<AppState, AppState>((state) => state);
+    useEffect(() => {
+        const resultsContainer = document.getElementById('results-container');
+        if (!resultsContainer) {
+            return undefined;
+        }
+        //tree-item element height is ~100px, using 50px here to be on the safe side.
+        const setPageSizeByContainerHeight = (pxHeight: number): void => {
+            actions.setPageSize(Math.ceil(pxHeight / 50));
+        };
+        const resizeObserver = new ResizeObserver((entries: ResizeObserverEntry[]) => {
+            const containerEntry = entries.find((entry) => entry?.target?.id === 'results-container');
+            if (containerEntry) {
+                setPageSizeByContainerHeight(containerEntry.contentRect.height);
+            }
+        });
+        // Set initial page size
+        setPageSizeByContainerHeight(resultsContainer.clientHeight);
+        resizeObserver.observe(resultsContainer);
+        return () => {
+            if (resizeObserver) {
+                resizeObserver.unobserve(resultsContainer);
+            }
+        };
+    }, []);
+
+    const fetchData = () => {
+        if (appState.guidedAnswerTreeSearchResult.resultSize > appState.pageSize) {
+            actions.searchTree({
+                query: appState.query,
+                filters: {
+                    product: appState.selectedProductFilters,
+                    component: appState.selectedComponentFilters
+                },
+                paging: {
+                    responseSize: appState.pageSize,
+                    offset: appState.guidedAnswerTreeSearchResult.trees.length
+                }
+            });
+        }
+    };
 
     let content;
     if (appState.loading) {
@@ -35,48 +76,57 @@ export function App(): ReactElement {
                 <FocusZone direction={FocusZoneDirection.bidirectional} isCircularNavigation={true}>
                     <FiltersRibbon />
                     <ul className="striped-list" role="listbox">
-                        {appState.guidedAnswerTreeSearchResult.trees.map((tree, index) => {
-                            return (
-                                <li key={`tree-item-${index}`} className="tree-item" role="option">
-                                    <button
-                                        className="guided-answer__tree"
-                                        onClick={(): void => {
-                                            actions.setActiveTree(tree);
-                                            actions.selectNode(tree.FIRST_NODE_ID);
-                                            document.body.focus();
-                                        }}>
-                                        <div className="guided-answer__tree__ul">
-                                            <h3
-                                                className="guided-answer__tree__title"
-                                                style={{ marginTop: tree.DESCRIPTION ? '0' : '10px' }}>
-                                                {tree.TITLE}
-                                            </h3>
-                                            <div className="bottom-section">
-                                                {tree.DESCRIPTION && (
-                                                    <span className="guided-answer__tree__desc">
-                                                        {tree.DESCRIPTION}
-                                                    </span>
-                                                )}
-                                                <div className="component-and-product-container">
-                                                    {tree.PRODUCT && (
-                                                        <div className="guided-answer__tree__product">
-                                                            <span className="bottom-title">Product: </span>
-                                                            {tree.PRODUCT.split(',')[0].trim()}
-                                                        </div>
+                        <InfiniteScroll
+                            dataLength={appState.guidedAnswerTreeSearchResult.trees.length} //This is important field to render the next data
+                            next={fetchData}
+                            loader={<VSCodeProgressRing id="loading-indicator" />}
+                            hasMore={
+                                appState.guidedAnswerTreeSearchResult.trees.length <
+                                appState.guidedAnswerTreeSearchResult.resultSize
+                            }>
+                            {appState.guidedAnswerTreeSearchResult.trees.map((tree, index) => {
+                                return (
+                                    <li key={`tree-item-${index}`} className="tree-item" role="option">
+                                        <button
+                                            className="guided-answer__tree"
+                                            onClick={(): void => {
+                                                actions.setActiveTree(tree);
+                                                actions.selectNode(tree.FIRST_NODE_ID);
+                                                document.body.focus();
+                                            }}>
+                                            <div className="guided-answer__tree__ul">
+                                                <h3
+                                                    className="guided-answer__tree__title"
+                                                    style={{ marginTop: tree.DESCRIPTION ? '0' : '10px' }}>
+                                                    {tree.TITLE}
+                                                </h3>
+                                                <div className="bottom-section">
+                                                    {tree.DESCRIPTION && (
+                                                        <span className="guided-answer__tree__desc">
+                                                            {tree.DESCRIPTION}
+                                                        </span>
                                                     )}
-                                                    {tree.COMPONENT && (
-                                                        <div className="guided-answer__tree__component">
-                                                            <span className="bottom-title">Component: </span>
-                                                            {tree.COMPONENT.split(',')[0].trim()}
-                                                        </div>
-                                                    )}
+                                                    <div className="component-and-product-container">
+                                                        {tree.PRODUCT && (
+                                                            <div className="guided-answer__tree__product">
+                                                                <span className="bottom-title">Product: </span>
+                                                                {tree.PRODUCT.split(',')[0].trim()}
+                                                            </div>
+                                                        )}
+                                                        {tree.COMPONENT && (
+                                                            <div className="guided-answer__tree__component">
+                                                                <span className="bottom-title">Component: </span>
+                                                                {tree.COMPONENT.split(',')[0].trim()}
+                                                            </div>
+                                                        )}
+                                                    </div>
                                                 </div>
                                             </div>
-                                        </div>
-                                    </button>
-                                </li>
-                            );
-                        })}
+                                        </button>
+                                    </li>
+                                );
+                            })}
+                        </InfiniteScroll>
                     </ul>
                 </FocusZone>
             );
@@ -89,7 +139,9 @@ export function App(): ReactElement {
                 showNavButons={appState.activeGuidedAnswerNode.length !== 0}
                 showSearch={appState.activeGuidedAnswerNode.length === 0}
             />
-            <main className="guided-answer__container">{content}</main>
+            <main className="guided-answer__container" id="results-container">
+                {content}
+            </main>
         </div>
     );
 }